# api/openapi.yaml
openapi: 3.1.0
info:
  title: SuperMonitor API
  version: 0.9.0
  description: Monitoring-only API. It intentionally exposes no model proxy endpoints.
servers:
  - url: /api/v1
security:
  - AdminSession: []
  - BearerAuth: []
paths:
...
  /auth/session:
    post:
      security: []
      operationId: createAuthenticationSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  minLength: 12
                  writeOnly: true
                  description: Administrator password containing at least 12 Unicode characters
                remember:
                  type: boolean
                  default: true
                  description: Keep the browser session for 30 days instead of 12 hours
      responses:
        "200":
          description: HttpOnly administrator session created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
        "401":
          description: Administrator password is invalid
    delete:
      operationId: deleteAuthenticationSession
      responses:
        "200":
          description: Browser administrator session cleared
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthStatus" }
  /auth/connect:
    post:
      security: []
      operationId: connectAuthenticationSession
      description: Browser form login for a user-selected SuperMonitor address. Passwords are exchanged for an HttpOnly session and never placed in a URL.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [address, password]
              properties:
                address: { type: string, format: uri }
                password: { type: string, minLength: 12, writeOnly: true }
                remember: { type: string, enum: ["1"] }
      responses:
        "303":
          description: Redirect to the console or back to the login gate with a non-sensitive error code
...
  /refresh:
    post:
      operationId: refreshAll
      responses:
        "202":
          description: Refresh completed for connected real accounts
        "409":
          description: A manual or scheduled account refresh is already running
...
  /providers/{providerID}/accounts/import:
    post:
      operationId: importProviderCredential
      parameters:
        - in: path
          name: providerID
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                alias: { type: string }
                file: { type: string, format: binary }
      responses:
        "201":
          description: Credential imported and live quota fetched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountSummary" }
  /providers/{providerID}/accounts/secret:
    post:
      operationId: connectProviderSecret
      parameters:
        - in: path
          name: providerID
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [secret]
              properties:
                alias: { type: string }
                secret: { type: string }
      responses:
        "201":
          description: Secret verified and live quota fetched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountSummary" }
  /providers/{providerID}/accounts/custom:
    post:
      operationId: connectCustomQuotaEndpoint
      parameters:
        - in: path
          name: providerID
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, method, authType, valuePath, label, unit, kind]
              properties:
                alias: { type: string }
                endpoint: { type: string, format: uri }
                method: { type: string, enum: [GET, POST] }
                authType: { type: string, enum: [bearer, header, cookie, none] }
                headerName: { type: string }
                headerPrefix: { type: string }
                secret: { type: string }
                requestBody: { type: string }
                valuePath: { type: string }
                totalPath: { type: string }
                resetAtPath: { type: string }
                expiresAtPath: { type: string }
                label: { type: string }
                unit: { type: string }
                kind: { type: string }
      responses:
        "201":
          description: Endpoint authenticated, response mapping verified, and account saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountSummary" }
  /providers/{providerID}/oauth:
    post:
      operationId: startProviderOAuth
      parameters:
        - in: path
          name: providerID
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                alias: { type: string }
      responses:
        "202":
          description: Official provider OAuth login started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceLoginSession" }
  /providers/{providerID}/oauth/{sessionID}:
    get:
      operationId: getProviderOAuth
      parameters:
        - in: path
          name: providerID
          required: true
          schema: { type: string }
        - in: path
          name: sessionID
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OAuth login status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceLoginSession" }
  /providers/kiro/oauth/callback:
    get:
      security: []
      operationId: completeKiroOAuth
      description: Public one-time OAuth callback protected by the server-held state and PKCE verifier.
      parameters:
        - in: query
          name: state
          required: true
          schema: { type: string }
        - in: query
          name: code
          schema: { type: string }
        - in: query
          name: error
          schema: { type: string }
        - in: query
          name: login_option
          schema: { type: string }
      responses:
        "200":
          description: Kiro authorization completed; the browser may close this page
          content:
            text/html:
              schema: { type: string }
        "400":
          description: Invalid, expired, denied, or failed Kiro authorization callback
          content:
            text/html:
              schema: { type: string }
  /providers/codex/accounts/import:
    post:
      operationId: importCodexAccount
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                alias: { type: string }
                file: { type: string, format: binary }
      responses:
        "201":
          description: Codex account imported and quota fetched
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountSummary" }
  /providers/codex/device-login:
    post:
      operationId: startCodexDeviceLogin
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                alias: { type: string }
      responses:
        "202":
          description: Official Codex device login started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceLoginSession" }
  /providers/codex/device-login/{sessionID}:
    get:
      operationId: getCodexDeviceLogin
      parameters:
        - in: path
          name: sessionID
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Device login status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceLoginSession" }
  /accounts/{accountID}/refresh:
    post:
      operationId: refreshAccount
      parameters:
        - in: path
          name: accountID
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Refreshed account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountSummary" }
  /accounts/{accountID}:
    delete:
      operationId: deleteAccount
      parameters:
        - in: path
          name: accountID
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Account, encrypted credential, and cached quota windows deleted
        "404":
          description: Account does not exist or was already deleted
  /notifications/channels:
    get:
      operationId: listNotificationChannels
      responses:
        "200":
          description: Notification channel summaries without credentials
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/NotificationChannel" }
    post:
      operationId: createNotificationChannel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationChannelInput" }
      responses:
        "201":
          description: Encrypted notification channel created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationChannel" }
  /notifications/channels/{channelID}:
    delete:
      operationId: deleteNotificationChannel
      parameters:
        - in: path
          name: channelID
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Notification channel and delivery receipts deleted
  /notifications/channels/{channelID}/test:
    post:
      operationId: testNotificationChannel
      parameters:
        - in: path
          name: channelID
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Test notification accepted by the configured channel
...
  /activities/{accountID}/{activityID}:
    post:
      operationId: runActivity
      parameters:
        - in: path
          name: accountID
          required: true
          schema: { type: string }
        - in: path
          name: activityID
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Activity executed and account quota refreshed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: supermonitor_admin
      description: Server-issued HttpOnly administrator session (12 hours or 30 days when remembered)
    BearerAuth:
      type: http
      scheme: bearer
      description: SUPMON_ADMIN_PASSWORD for API automation
  schemas:
    AuthStatus:
      type: object
      required: [required, authenticated]
      properties:
        required: { type: boolean }
        authenticated: { type: boolean }
...
    NotificationChannel:
      type: object
      required: [id, kind, name, target, enabled, createdAt, updatedAt]
      properties:
        id: { type: string }
        kind: { type: string, enum: [feishu, qq_mail] }
        name: { type: string }
        target: { type: string, description: Masked target summary; credentials are never returned }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    NotificationChannelInput:
      type: object
      required: [kind]
      properties:
        kind: { type: string, enum: [feishu, qq_mail] }
        name: { type: string }
        webhookUrl: { type: string, format: uri }
        sender: { type: string, format: email }
        authCode: { type: string, writeOnly: true }
        recipient: { type: string, format: email }
...
    AccountSummary:
      type: object
      required: [id, providerId, provider, region, alias, services, primaryMetric, secondaryMetric, status, source, lastRefreshedAt, nextRefreshAt, synthetic, quotaWindows]
      properties:
        id: { type: string }
        providerId: { type: string }
        provider: { type: string }
        region: { type: string }
        alias: { type: string }
        services:
          type: array
          items: { type: string }
        primaryMetric: { type: string }
        secondaryMetric: { type: string }
        status: { type: string }
        source: { type: string }
        lastRefreshedAt: { type: string, format: date-time }
        nextRefreshAt: { type: string, format: date-time }
        error: { type: string }
        email: { type: string }
        plan: { type: string }
        authMethod: { type: string }
        synthetic: { type: boolean }
        quotaWindows:
          type: array
          items: { $ref: "#/components/schemas/QuotaSignal" }
...
    DeviceLoginSession:
      type: object
      required: [id, provider, status, verifyUrl, expiresAt, message]
      properties:
        id: { type: string }
        provider: { type: string }
        status: { type: string, enum: [pending, completed, failed] }
        userCode: { type: string }
        verifyUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        accountId: { type: string }
        message: { type: string }